conditionally avoid -pgmlg++ on darwin
It appears that the flag `pgmlg++` was added so as to use g++ for
linking the `css` executable.  Unfortunately, on darwin, command-line
arguments such as `-no-pie` and `-stdlib=libc++` are passed to `g++`,
which is believed to be `clang++`.  Those flags trigger a warning in the
compilation phase, but also trigger an error in the linker phase, when
passed to the real `g++`.

This proposed change conditionally ignores this option.  I'm not sure
whether it is safe to do so, and what are the consequences, but the
executable `css` builds, and at least can be run with no input on
Darwin.  Not knowing the intent of the original flag, I cannot vouch
for whether this retains the expected functionality.